Forced SymbolTable's IDType parameter to be unsigned and made size_t the default.
Since code used it as an index and compared it with vector's size, negative
values weren't really an option. Now a static_assert makes this explicit.

Had to make some minor updates to the test suite, since singed integer types
were used in most places.

## Successful Builds

1. On Linux Mint 18 Sarah (Linux base 4.4.0-21-generic x86_64)

    1. Using clang @ 3.8.0-2ubuntu4 (tags/RELEASE_380/final)
    2. Using gcc @ 5.4.0 20160609 (Ubuntu 5.4.0-6ubuntu1~16.04.9)

   With the following packages installed:

    1. libglew-dev @ 1.13.0-2
    2. freeglut3 @ 2.8.1-2
    3. libeigen3-dev @ 3.3~beta1-2
    4. libboost-all-dev @ 1.58.0.1ubuntu1
    5. cmake @ 3.5.1-1ubuntu3

2. On Windows 7 Professional 64-bit Service Pack 1

    1. Using Visual Studio 2017 Community Edition
        a. With VC++ 2017 v141 toolset (x86,x64)

   With the following applications installed:

    1. CMake 3.11.1 (64-bit)
    2. vcpkg 0.0.110-2018-04-28-a9d989ec056246ea320a56ca7d02b161a6fee387

   With the following vcpkg packages installed:

    1. glew @ 2.1.0-1 (x64-windows)
    2. freeglut @ 3.0.0-3 (x64-windows)
    3. eigen3 @ 3.3.4-2 (x64-windows)
    4. boost @ 1.67.0 (x64-windows)

3. On Windows 8.1 Pro 64-bit

    1. Using Visual Studio 2017 Community Edition
        a. With VC++ 2017 v141 toolset (x86,x64)

   With the following applications installed:

    1. CMake 3.11.1 (64-bit)
    2. vcpkg 0.0.110-2018-04-28-a9d989ec056246ea320a56ca7d02b161a6fee387

   With the following vcpkg packages installed:

    1. glew @ 2.1.0-1 (x64-windows)
    2. freeglut @ 3.0.0-3 (x64-windows)
    3. eigen3 @ 3.3.4-2 (x64-windows)
    4. boost @ 1.67.0 (x64-windows)

--HG--
branch : enhancement

+++ source/symbolic computation.hpp
@@ -184,7 +184,7 @@ namespace Symbolic
 		 *	Signed, rational and fixed point real numbers are represented using negation,
 		 *	division and natural numbers.
 		 */
-		template<typename UIntType = unsigned long long int, typename NameType = std::string, typename IDType = int>
+		template<typename UIntType = unsigned long long int, typename NameType = std::string, typename IDType = size_t>
 		class Expression
 		{
 			/**********************
@@ -896,7 +896,7 @@ namespace Symbolic
 		template<typename UIntType, typename NameType>
 		struct ArgType<Primitive::Constant,UIntType,NameType>{using type = UIntType;};
 
-		template<Primitive primitive, typename UIntType = unsigned long long int, typename NameType = std::string, typename IDType = int>
+		template<Primitive primitive, typename UIntType = unsigned long long int, typename NameType = std::string, typename IDType = size_t>
 		struct ExpressionBuilder
 		{
 			// Member Types
